map alignment| scanning fix
map alignment| scanning fix

+++ Assets/Scripts/Models/Map.cs
@@ -3,8 +3,8 @@ using UnityEngine;
 
 public class Map : MonoBehaviour
 {
-    public double longitude { get; set; }
-    public double latitude { get; set; }
+    public float longitude { get; set; }
+    public float latitude { get; set; }
 
     public string name { get; set; }  
     public List<Anchor> anchors { get; set; }
